# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/menu_category": {
            "get": {
                "tags": [
                    "Menu category"
                ],
                "summary": "Вывод списка категорий меню",
                "operationId": "574f78f83bb91f90dc20b6760cc747f4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Страница",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Наименование",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Завтрак"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Направление сортировки",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Завтрак"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "Images/MenuCategory/S9EXsmeHwGPkewmzcm0f6T9RtLHiYjzjvEj9z1Tx.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menu category"
                ],
                "summary": "Ввод новой категории меню",
                "operationId": "6619c8f3c6e9d2ed6596c0328f19a1db",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Завтрак"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Завтрак"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "Images/MenuCategory/S9EXsmeHwGPkewmzcm0f6T9RtLHiYjzjvEj9z1Tx.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}